fix(api): make wavelog auth/station_info accept GridTracker's request shape
GridTracker's Cloudlog client (inspected via app.asar src) sends
requests that don't quite match what nextlog implemented:

  1. POST {URL}/index.php/api/auth/{KEY}       (not GET)
  2. POST {URL}/index.php/api/station_info/{KEY}   (key in path, no body)

Our /api/auth/[key] only exported GET → Vercel returned 405 Method Not
Allowed with an empty body → GT logged "Invalid Response" and stalled
at "Testing API Key". Our /api/station_info only accepted the key in
the JSON body, so /api/station_info/{KEY} 404'd → after auth's "OK"
the profile-fill response was empty → GT overwrote "OK" with the same
"Invalid Response" message.

This commit:

- /api/auth/[key]: alias GET and POST to one handler; add an OPTIONS
  handler for CORS preflight. Matches wavelog/CodeIgniter behavior
  where controllers respond to any HTTP method.
- /api/station_info/[key]: new dynamic route, POST + GET both supported
- /api/station_info (body form) preserved as-is, no breaking change
- Extract shared lookup logic into src/lib/station-info.ts so both
  routes return identical wavelog-shape JSON

Verified GT's full flow by reading
  C:\Users\burns\AppData\Local\Programs\GridTracker2\resources\app.asar
  src/renderer/lib/adif.js — functions CloudlogTest,
  CloudlogGetProfiles, CloudlogTestApiKey, CloudlogFillProfiles.

After this lands, GT's Test button should show "OK" and populate the
station profile dropdown.
