+++ src/app/api/auth/[key]/route.ts
@@ -1,7 +1,12 @@
-// GET /api/auth/<key> — wavelog wire-compatible API key validation.
+// /api/auth/<key> — wavelog wire-compatible API key validation.
 // Returns XML, matching wavelog's <auth>...</auth> shape. The HTTP status is
 // 200 regardless of whether the key is valid (matching wavelog's quirk —
 // clients distinguish via the body, not the status).
+//
+// Accepts both GET and POST: wavelog runs on CodeIgniter, whose controllers
+// answer any HTTP method by default, and several clients (GridTracker, in
+// particular) POST to this URL even though the key is in the path. Honor
+// both verbs so those clients work.
 
 import { NextRequest } from 'next/server';
 import { verifyApiKeyValue } from '@/lib/api-auth';
@@ -22,7 +27,7 @@ function invalidKeyResponse() {
   return xmlResponse('<auth>\n  <message>Key Invalid - either not found or disabled</message>\n</auth>');
 }
 
-export async function GET(
+async function handle(
   _request: NextRequest,
   { params }: { params: Promise<{ key: string }> },
 ) {
@@ -34,4 +39,19 @@ export async function GET(
 
   const rights = authResult.auth.isReadOnly ? 'r' : 'rw';
   return xmlResponse(`<auth>\n  <status>Valid</status>\n  <rights>${rights}</rights>\n</auth>`);
+}
+
+export const GET = handle;
+export const POST = handle;
+
+export function OPTIONS() {
+  return new Response(null, {
+    status: 200,
+    headers: {
+      'Access-Control-Allow-Origin': '*',
+      'Access-Control-Allow-Methods': 'GET, POST, OPTIONS',
+      'Access-Control-Allow-Headers': 'Content-Type, Authorization, X-API-Key',
+      'Access-Control-Max-Age': '86400',
+    },
+  });
 }
